Fix bug when promises are not resolved
When a user clicks on a new lecture in the SPA,
the page may or may not contain a video. If the
page initially doesn't have a video, the script
will keep searching for it. However, if the user
switches to a new lecture before the video is
found, the previous search might still resolve
with the current video element, which could
interfere with the new lecture. To handle this, I
compare the ID of the current lecture with the ID
associated with the element being searched for. If
the IDs don't match, I resolve the previous
promise as null to ensure it doesn't mistakenly
affect the new lecture.